chore: move main bgImg to the public folder
